move over wellContentsAllLabware tests to the selectors new location
